Use uint64 in countQuery, TxCount, Count()

+++ find_test.go
@@ -1168,7 +1168,7 @@ func TestCount(t *testing.T) {
 					t.Fatalf("Error counting data from badgerhold: %s", err)
 				}
 
-				if count != len(tst.result) {
+				if count != uint64(len(tst.result)) {
 					t.Fatalf("Count result is %d wanted %d.", count, len(tst.result))
 				}
 			})
